cleanup: refactor getTagListSections to TagsOptionsListUtils
