+++ src/components/TagPicker/index.tsx
@@ -7,18 +7,11 @@ import useThemeStyles from '@hooks/useThemeStyles';
 import * as OptionsListUtils from '@libs/OptionsListUtils';
 import * as PolicyUtils from '@libs/PolicyUtils';
 import type * as ReportUtils from '@libs/ReportUtils';
+import type {SelectedTagOption} from '@libs/TagsOptionsListUtils';
+import * as TagOptionListUtils from '@libs/TagsOptionsListUtils';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type {PolicyTag, PolicyTags} from '@src/types/onyx';
-import type {PendingAction} from '@src/types/onyx/OnyxCommon';
-
-type SelectedTagOption = {
-    name: string;
-    enabled: boolean;
-    isSelected?: boolean;
-    accountID: number | undefined;
-    pendingAction?: PendingAction;
-};
 
 type TagPickerProps = {
     /** The policyID we are getting tags for */
@@ -81,15 +74,12 @@ function TagPicker({selectedTag, tagListName, policyID, tagListIndex, shouldShow
 
     const sections = useMemo(
         () =>
-            OptionsListUtils.getFilteredOptions({
+            TagOptionListUtils.getTagListSections({
                 searchValue,
                 selectedOptions,
-                includeP2P: false,
-                includeTags: true,
                 tags: enabledTags,
                 recentlyUsedTags: policyRecentlyUsedTagsList,
-                canInviteUser: false,
-            }).tagOptions,
+            }),
         [searchValue, enabledTags, selectedOptions, policyRecentlyUsedTagsList],
     );
