+++ src/libs/OptionsListUtils.ts
@@ -9,7 +9,6 @@ import Onyx from 'react-native-onyx';
 import type {OnyxCollection, OnyxEntry} from 'react-native-onyx';
 import type {SetNonNullable} from 'type-fest';
 import {FallbackAvatar} from '@components/Icon/Expensicons';
-import type {SelectedTagOption} from '@components/TagPicker';
 import type {IOUAction} from '@src/CONST';
 import CONST from '@src/CONST';
 import type {TranslationPaths} from '@src/languages/types';
@@ -24,8 +23,6 @@ import type {
     PolicyCategories,
     PolicyCategory,
     PolicyTag,
-    PolicyTagLists,
-    PolicyTags,
     Report,
     ReportAction,
     ReportActions,
@@ -42,7 +39,6 @@ import {isEmptyObject} from '@src/types/utils/EmptyObject';
 import times from '@src/utils/times';
 import Timing from './actions/Timing';
 import filterArrayByMatch from './filterArrayByMatch';
-import localeCompare from './LocaleCompare';
 import * as LocalePhoneNumber from './LocalePhoneNumber';
 import * as Localize from './Localize';
 import * as LoginUtils from './LoginUtils';
@@ -162,9 +158,6 @@ type GetOptionsConfig = {
     includeCategories?: boolean;
     categories?: PolicyCategories;
     recentlyUsedCategories?: string[];
-    includeTags?: boolean;
-    tags?: PolicyTags | Array<SelectedTagOption | PolicyTag>;
-    recentlyUsedTags?: string[];
     canInviteUser?: boolean;
     includeSelectedOptions?: boolean;
     includeTaxRates?: boolean;
@@ -215,7 +208,6 @@ type Options = {
     userToInvite: ReportUtils.OptionData | null;
     currentUserOption: ReportUtils.OptionData | null | undefined;
     categoryOptions: CategoryTreeSection[];
-    tagOptions: CategorySection[];
     taxRatesOptions: CategorySection[];
     policyReportFieldOptions?: CategorySection[] | null;
 };
@@ -983,16 +975,6 @@ function sortCategories(categories: Record<string, Category>): Category[] {
 }
 
 /**
- * Sorts tags alphabetically by name.
- */
-function sortTags(tags: Record<string, PolicyTag | SelectedTagOption> | Array<PolicyTag | SelectedTagOption>) {
-    const sortedTags = Array.isArray(tags) ? tags : Object.values(tags);
-
-    // Use lodash's sortBy to ensure consistency with oldDot.
-    return lodashSortBy(sortedTags, 'name', localeCompare);
-}
-
-/**
  * Builds the options for the category tree hierarchy via indents
  *
  * @param options - an initial object array
@@ -1172,138 +1154,6 @@ function getCategoryListSections(
     });
 
     return categorySections;
-}
-
-/**
- * Transforms the provided tags into option objects.
- *
- * @param tags - an initial tag array
- */
-function getTagsOptions(tags: Array<Pick<PolicyTag, 'name' | 'enabled' | 'pendingAction'>>, selectedOptions?: SelectedTagOption[]): Option[] {
-    return tags.map((tag) => {
-        // This is to remove unnecessary escaping backslash in tag name sent from backend.
-        const cleanedName = PolicyUtils.getCleanedTagName(tag.name);
-        return {
-            text: cleanedName,
-            keyForList: tag.name,
-            searchText: tag.name,
-            tooltipText: cleanedName,
-            isDisabled: !tag.enabled || tag.pendingAction === CONST.RED_BRICK_ROAD_PENDING_ACTION.DELETE,
-            isSelected: selectedOptions?.some((selectedTag) => selectedTag.name === tag.name),
-            pendingAction: tag.pendingAction,
-        };
-    });
-}
-
-/**
- * Build the section list for tags
- */
-function getTagListSections(
-    tags: Array<PolicyTag | SelectedTagOption>,
-    recentlyUsedTags: string[],
-    selectedOptions: SelectedTagOption[],
-    searchInputValue: string,
-    maxRecentReportsToShow: number,
-) {
-    const tagSections = [];
-    const sortedTags = sortTags(tags) as PolicyTag[];
-    const selectedOptionNames = selectedOptions.map((selectedOption) => selectedOption.name);
-    const enabledTags = sortedTags.filter((tag) => tag.enabled);
-    const enabledTagsNames = enabledTags.map((tag) => tag.name);
-    const enabledTagsWithoutSelectedOptions = enabledTags.filter((tag) => !selectedOptionNames.includes(tag.name));
-    const selectedTagsWithDisabledState: SelectedTagOption[] = [];
-    const numberOfTags = enabledTags.length;
-
-    selectedOptions.forEach((tag) => {
-        if (enabledTagsNames.includes(tag.name)) {
-            selectedTagsWithDisabledState.push({...tag, enabled: true});
-            return;
-        }
-        selectedTagsWithDisabledState.push({...tag, enabled: false});
-    });
-
-    // If all tags are disabled but there's a previously selected tag, show only the selected tag
-    if (numberOfTags === 0 && selectedOptions.length > 0) {
-        tagSections.push({
-            // "Selected" section
-            title: '',
-            shouldShow: false,
-            data: getTagsOptions(selectedTagsWithDisabledState, selectedOptions),
-        });
-
-        return tagSections;
-    }
-
-    if (searchInputValue) {
-        const enabledSearchTags = enabledTagsWithoutSelectedOptions.filter((tag) => PolicyUtils.getCleanedTagName(tag.name.toLowerCase()).includes(searchInputValue.toLowerCase()));
-        const selectedSearchTags = selectedTagsWithDisabledState.filter((tag) => PolicyUtils.getCleanedTagName(tag.name.toLowerCase()).includes(searchInputValue.toLowerCase()));
-        const tagsForSearch = [...selectedSearchTags, ...enabledSearchTags];
-
-        tagSections.push({
-            // "Search" section
-            title: '',
-            shouldShow: true,
-            data: getTagsOptions(tagsForSearch, selectedOptions),
-        });
-
-        return tagSections;
-    }
-
-    if (numberOfTags < CONST.TAG_LIST_THRESHOLD) {
-        tagSections.push({
-            // "All" section when items amount less than the threshold
-            title: '',
-            shouldShow: false,
-            data: getTagsOptions([...selectedTagsWithDisabledState, ...enabledTagsWithoutSelectedOptions], selectedOptions),
-        });
-
-        return tagSections;
-    }
-
-    const filteredRecentlyUsedTags = recentlyUsedTags
-        .filter((recentlyUsedTag) => {
-            const tagObject = tags.find((tag) => tag.name === recentlyUsedTag);
-            return !!tagObject?.enabled && !selectedOptionNames.includes(recentlyUsedTag) && tagObject?.pendingAction !== CONST.RED_BRICK_ROAD_PENDING_ACTION.DELETE;
-        })
-        .map((tag) => ({name: tag, enabled: true}));
-
-    if (selectedOptions.length) {
-        tagSections.push({
-            // "Selected" section
-            title: '',
-            shouldShow: true,
-            data: getTagsOptions(selectedTagsWithDisabledState, selectedOptions),
-        });
-    }
-
-    if (filteredRecentlyUsedTags.length > 0) {
-        const cutRecentlyUsedTags = filteredRecentlyUsedTags.slice(0, maxRecentReportsToShow);
-
-        tagSections.push({
-            // "Recent" section
-            title: Localize.translateLocal('common.recent'),
-            shouldShow: true,
-            data: getTagsOptions(cutRecentlyUsedTags, selectedOptions),
-        });
-    }
-
-    tagSections.push({
-        // "All" section when items amount more than the threshold
-        title: Localize.translateLocal('common.all'),
-        shouldShow: true,
-        data: getTagsOptions(enabledTagsWithoutSelectedOptions, selectedOptions),
-    });
-
-    return tagSections;
-}
-
-/**
- * Verifies that there is at least one enabled tag
- */
-function hasEnabledTags(policyTagList: Array<PolicyTagLists[keyof PolicyTagLists]>) {
-    const policyTagValueList = policyTagList.map(({tags}) => Object.values(tags)).flat();
-
-    return hasEnabledOptions(policyTagValueList);
 }
 
 /**
@@ -1714,9 +1564,6 @@ function getOptions(
         includeCategories = false,
         categories = {},
         recentlyUsedCategories = [],
-        includeTags = false,
-        tags = {},
-        recentlyUsedTags = [],
         canInviteUser = true,
         includeSelectedOptions = false,
         transactionViolations = {},
@@ -1743,21 +1590,6 @@ function getOptions(
             userToInvite: null,
             currentUserOption: null,
             categoryOptions,
-            tagOptions: [],
-            taxRatesOptions: [],
-        };
-    }
-
-    if (includeTags) {
-        const tagOptions = getTagListSections(Object.values(tags), recentlyUsedTags, selectedOptions as SelectedTagOption[], searchInputValue, maxRecentReportsToShow);
-
-        return {
-            recentReports: [],
-            personalDetails: [],
-            userToInvite: null,
-            currentUserOption: null,
-            categoryOptions: [],
-            tagOptions,
             taxRatesOptions: [],
         };
     }
@@ -1771,7 +1603,6 @@ function getOptions(
             userToInvite: null,
             currentUserOption: null,
             categoryOptions: [],
-            tagOptions: [],
             taxRatesOptions,
         };
     }
@@ -1784,7 +1615,6 @@ function getOptions(
             userToInvite: null,
             currentUserOption: null,
             categoryOptions: [],
-            tagOptions: [],
             taxRatesOptions: [],
             policyReportFieldOptions: transformedPolicyReportFieldOptions,
         };
@@ -2046,7 +1876,6 @@ function getOptions(
         userToInvite: canInviteUser ? userToInvite : null,
         currentUserOption,
         categoryOptions: [],
-        tagOptions: [],
         taxRatesOptions: [],
     };
 }
@@ -2133,9 +1962,6 @@ type FilteredOptionsParams = {
     includeCategories?: boolean;
     categories?: PolicyCategories;
     recentlyUsedCategories?: string[];
-    includeTags?: boolean;
-    tags?: PolicyTags | Array<PolicyTag | SelectedTagOption>;
-    recentlyUsedTags?: string[];
     canInviteUser?: boolean;
     includeSelectedOptions?: boolean;
     includeTaxRates?: boolean;
@@ -2174,9 +2000,6 @@ function getFilteredOptions(params: FilteredOptionsParamsWithDefaultSearchValue 
         includeCategories = false,
         categories = {},
         recentlyUsedCategories = [],
-        includeTags = false,
-        tags = {},
-        recentlyUsedTags = [],
         canInviteUser = true,
         includeSelectedOptions = false,
         includeTaxRates = false,
@@ -2204,9 +2027,6 @@ function getFilteredOptions(params: FilteredOptionsParamsWithDefaultSearchValue 
             includeCategories,
             categories,
             recentlyUsedCategories,
-            includeTags,
-            tags,
-            recentlyUsedTags,
             canInviteUser,
             includeSelectedOptions,
             includeTaxRates,
@@ -2248,9 +2068,6 @@ function getAttendeeOptions(
             includeCategories: false,
             categories: {},
             recentlyUsedCategories: [],
-            includeTags: false,
-            tags: {},
-            recentlyUsedTags: [],
             canInviteUser,
             includeSelectedOptions: false,
             includeTaxRates: false,
@@ -2550,7 +2367,6 @@ function filterOptions(options: Options, searchInputValue: string, config?: Filt
             userToInvite: null,
             currentUserOption,
             categoryOptions: [],
-            tagOptions: [],
             taxRatesOptions: [],
         };
     }, options);
@@ -2586,7 +2402,6 @@ function filterOptions(options: Options, searchInputValue: string, config?: Filt
         userToInvite,
         currentUserOption: matchResults.currentUserOption,
         categoryOptions: [],
-        tagOptions: [],
         taxRatesOptions: [],
     };
 }
@@ -2602,7 +2417,6 @@ function getEmptyOptions(): Options {
         userToInvite: null,
         currentUserOption: null,
         categoryOptions: [],
-        tagOptions: [],
         taxRatesOptions: [],
     };
 }
@@ -2637,9 +2451,7 @@ export {
     hasEnabledOptions,
     sortCategories,
     sortAlphabetically,
-    sortTags,
     getCategoryOptionTree,
-    hasEnabledTags,
     formatMemberForList,
     formatSectionsFromSearchTerm,
     getShareLogOptions,
