+++ tests/unit/TagsOptionsListUtilsTest.ts
@@ -1,0 +1,543 @@
+import type * as OptionsListUtils from '@libs/OptionsListUtils';
+import type {SelectedTagOption} from '@libs/TagsOptionsListUtils';
+import * as TagsOptionsListUtils from '@libs/TagsOptionsListUtils';
+
+describe('TagsOptionsListUtils', () => {
+    it('getTagListSections()', () => {
+        const search = 'ing';
+        const emptySearch = '';
+        const wrongSearch = 'bla bla';
+        const recentlyUsedTags = ['Engineering', 'HR'];
+
+        const selectedOptions: SelectedTagOption[] = [
+            {
+                name: 'Medical',
+                enabled: true,
+                accountID: undefined,
+            },
+        ];
+        const smallTagsList: Record<string, SelectedTagOption> = {
+            Engineering: {
+                enabled: false,
+                name: 'Engineering',
+                accountID: undefined,
+            },
+            Medical: {
+                enabled: true,
+                name: 'Medical',
+                accountID: undefined,
+            },
+            Accounting: {
+                enabled: true,
+                name: 'Accounting',
+                accountID: undefined,
+            },
+            HR: {
+                enabled: true,
+                name: 'HR',
+                accountID: undefined,
+                pendingAction: 'delete',
+            },
+        };
+        const smallResultList: OptionsListUtils.CategorySection[] = [
+            {
+                title: '',
+                shouldShow: false,
+                // data sorted alphabetically by name
+                data: [
+                    {
+                        text: 'Accounting',
+                        keyForList: 'Accounting',
+                        searchText: 'Accounting',
+                        tooltipText: 'Accounting',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                    {
+                        text: 'HR',
+                        keyForList: 'HR',
+                        searchText: 'HR',
+                        tooltipText: 'HR',
+                        isDisabled: true,
+                        isSelected: false,
+                        pendingAction: 'delete',
+                    },
+                    {
+                        text: 'Medical',
+                        keyForList: 'Medical',
+                        searchText: 'Medical',
+                        tooltipText: 'Medical',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                ],
+            },
+        ];
+        const smallSearchResultList: OptionsListUtils.CategorySection[] = [
+            {
+                title: '',
+                shouldShow: true,
+                data: [
+                    {
+                        text: 'Accounting',
+                        keyForList: 'Accounting',
+                        searchText: 'Accounting',
+                        tooltipText: 'Accounting',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                ],
+            },
+        ];
+        const smallWrongSearchResultList: OptionsListUtils.CategoryTreeSection[] = [
+            {
+                title: '',
+                shouldShow: true,
+                data: [],
+            },
+        ];
+        const largeTagsList: Record<string, SelectedTagOption> = {
+            Engineering: {
+                enabled: false,
+                name: 'Engineering',
+                accountID: undefined,
+            },
+            Medical: {
+                enabled: true,
+                name: 'Medical',
+                accountID: undefined,
+            },
+            Accounting: {
+                enabled: true,
+                name: 'Accounting',
+                accountID: undefined,
+            },
+            HR: {
+                enabled: true,
+                name: 'HR',
+                accountID: undefined,
+            },
+            Food: {
+                enabled: true,
+                name: 'Food',
+                accountID: undefined,
+            },
+            Traveling: {
+                enabled: false,
+                name: 'Traveling',
+                accountID: undefined,
+            },
+            Cleaning: {
+                enabled: true,
+                name: 'Cleaning',
+                accountID: undefined,
+            },
+            Software: {
+                enabled: true,
+                name: 'Software',
+                accountID: undefined,
+            },
+            OfficeSupplies: {
+                enabled: false,
+                name: 'Office Supplies',
+                accountID: undefined,
+            },
+            Taxes: {
+                enabled: true,
+                name: 'Taxes',
+                accountID: undefined,
+                pendingAction: 'delete',
+            },
+            Benefits: {
+                enabled: true,
+                name: 'Benefits',
+                accountID: undefined,
+            },
+        };
+        const largeResultList: OptionsListUtils.CategorySection[] = [
+            {
+                title: '',
+                shouldShow: true,
+                data: [
+                    {
+                        text: 'Medical',
+                        keyForList: 'Medical',
+                        searchText: 'Medical',
+                        tooltipText: 'Medical',
+                        isDisabled: false,
+                        isSelected: true,
+                        pendingAction: undefined,
+                    },
+                ],
+            },
+            {
+                title: 'Recent',
+                shouldShow: true,
+                data: [
+                    {
+                        text: 'HR',
+                        keyForList: 'HR',
+                        searchText: 'HR',
+                        tooltipText: 'HR',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                ],
+            },
+            {
+                title: 'All',
+                shouldShow: true,
+                // data sorted alphabetically by name
+                data: [
+                    {
+                        text: 'Accounting',
+                        keyForList: 'Accounting',
+                        searchText: 'Accounting',
+                        tooltipText: 'Accounting',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                    {
+                        text: 'Benefits',
+                        keyForList: 'Benefits',
+                        searchText: 'Benefits',
+                        tooltipText: 'Benefits',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                    {
+                        text: 'Cleaning',
+                        keyForList: 'Cleaning',
+                        searchText: 'Cleaning',
+                        tooltipText: 'Cleaning',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                    {
+                        text: 'Food',
+                        keyForList: 'Food',
+                        searchText: 'Food',
+                        tooltipText: 'Food',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                    {
+                        text: 'HR',
+                        keyForList: 'HR',
+                        searchText: 'HR',
+                        tooltipText: 'HR',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                    {
+                        text: 'Software',
+                        keyForList: 'Software',
+                        searchText: 'Software',
+                        tooltipText: 'Software',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                    {
+                        text: 'Taxes',
+                        keyForList: 'Taxes',
+                        searchText: 'Taxes',
+                        tooltipText: 'Taxes',
+                        isDisabled: true,
+                        isSelected: false,
+                        pendingAction: 'delete',
+                    },
+                ],
+            },
+        ];
+        const largeSearchResultList: OptionsListUtils.CategorySection[] = [
+            {
+                title: '',
+                shouldShow: true,
+                data: [
+                    {
+                        text: 'Accounting',
+                        keyForList: 'Accounting',
+                        searchText: 'Accounting',
+                        tooltipText: 'Accounting',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                    {
+                        text: 'Cleaning',
+                        keyForList: 'Cleaning',
+                        searchText: 'Cleaning',
+                        tooltipText: 'Cleaning',
+                        isDisabled: false,
+                        isSelected: false,
+                        pendingAction: undefined,
+                    },
+                ],
+            },
+        ];
+        const largeWrongSearchResultList: OptionsListUtils.CategoryTreeSection[] = [
+            {
+                title: '',
+                shouldShow: true,
+                data: [],
+            },
+        ];
+
+        const smallResult = TagsOptionsListUtils.getTagListSections({searchValue: emptySearch, tags: smallTagsList});
+        expect(smallResult).toStrictEqual(smallResultList);
+
+        const smallSearchResult = TagsOptionsListUtils.getTagListSections({searchValue: search, tags: smallTagsList});
+        expect(smallSearchResult).toStrictEqual(smallSearchResultList);
+
+        const smallWrongSearchResult = TagsOptionsListUtils.getTagListSections({searchValue: wrongSearch, tags: smallTagsList});
+        expect(smallWrongSearchResult).toStrictEqual(smallWrongSearchResultList);
+
+        const largeResult = TagsOptionsListUtils.getTagListSections({searchValue: emptySearch, selectedOptions, tags: largeTagsList, recentlyUsedTags});
+        expect(largeResult).toStrictEqual(largeResultList);
+
+        const largeSearchResult = TagsOptionsListUtils.getTagListSections({searchValue: search, selectedOptions, tags: largeTagsList, recentlyUsedTags});
+        expect(largeSearchResult).toStrictEqual(largeSearchResultList);
+
+        const largeWrongSearchResult = TagsOptionsListUtils.getTagListSections({
+            searchValue: wrongSearch,
+            selectedOptions,
+            tags: largeTagsList,
+            recentlyUsedTags,
+        });
+        expect(largeWrongSearchResult).toStrictEqual(largeWrongSearchResultList);
+    });
+
+    it('sortTags', () => {
+        const createTagObjects = (names: string[]) => names.map((name) => ({name, enabled: true}));
+
+        const unorderedTagNames = ['10bc', 'b', '0a', '1', '中国', 'b10', '!', '2', '0', '@', 'a1', 'a', '3', 'b1', '日本', '$', '20', '20a', '#', 'a20', 'c', '10'];
+        const expectedOrderNames = ['!', '#', '$', '0', '0a', '1', '10', '10bc', '2', '20', '20a', '3', '@', 'a', 'a1', 'a20', 'b', 'b1', 'b10', 'c', '中国', '日本'];
+        const unorderedTags = createTagObjects(unorderedTagNames);
+        const expectedOrder = createTagObjects(expectedOrderNames);
+        expect(TagsOptionsListUtils.sortTags(unorderedTags)).toStrictEqual(expectedOrder);
+
+        const unorderedTagNames2 = ['0', 'a1', '1', 'b1', '3', '10', 'b10', 'a', '2', 'c', '20', 'a20', 'b'];
+        const expectedOrderNames2 = ['0', '1', '10', '2', '20', '3', 'a', 'a1', 'a20', 'b', 'b1', 'b10', 'c'];
+        const unorderedTags2 = createTagObjects(unorderedTagNames2);
+        const expectedOrder2 = createTagObjects(expectedOrderNames2);
+        expect(TagsOptionsListUtils.sortTags(unorderedTags2)).toStrictEqual(expectedOrder2);
+
+        const unorderedTagNames3 = [
+            '61',
+            '39',
+            '97',
+            '93',
+            '77',
+            '71',
+            '22',
+            '27',
+            '30',
+            '64',
+            '91',
+            '24',
+            '33',
+            '60',
+            '21',
+            '85',
+            '59',
+            '76',
+            '42',
+            '67',
+            '13',
+            '96',
+            '84',
+            '44',
+            '68',
+            '31',
+            '62',
+            '87',
+            '50',
+            '4',
+            '100',
+            '12',
+            '28',
+            '49',
+            '53',
+            '5',
+            '45',
+            '14',
+            '55',
+            '78',
+            '11',
+            '35',
+            '75',
+            '18',
+            '9',
+            '80',
+            '54',
+            '2',
+            '34',
+            '48',
+            '81',
+            '6',
+            '73',
+            '15',
+            '98',
+            '25',
+            '8',
+            '99',
+            '17',
+            '90',
+            '47',
+            '1',
+            '10',
+            '38',
+            '66',
+            '57',
+            '23',
+            '86',
+            '29',
+            '3',
+            '65',
+            '74',
+            '19',
+            '56',
+            '63',
+            '20',
+            '7',
+            '32',
+            '46',
+            '70',
+            '26',
+            '16',
+            '83',
+            '37',
+            '58',
+            '43',
+            '36',
+            '69',
+            '79',
+            '72',
+            '41',
+            '94',
+            '95',
+            '82',
+            '51',
+            '52',
+            '89',
+            '88',
+            '40',
+            '92',
+        ];
+        const expectedOrderNames3 = [
+            '1',
+            '10',
+            '100',
+            '11',
+            '12',
+            '13',
+            '14',
+            '15',
+            '16',
+            '17',
+            '18',
+            '19',
+            '2',
+            '20',
+            '21',
+            '22',
+            '23',
+            '24',
+            '25',
+            '26',
+            '27',
+            '28',
+            '29',
+            '3',
+            '30',
+            '31',
+            '32',
+            '33',
+            '34',
+            '35',
+            '36',
+            '37',
+            '38',
+            '39',
+            '4',
+            '40',
+            '41',
+            '42',
+            '43',
+            '44',
+            '45',
+            '46',
+            '47',
+            '48',
+            '49',
+            '5',
+            '50',
+            '51',
+            '52',
+            '53',
+            '54',
+            '55',
+            '56',
+            '57',
+            '58',
+            '59',
+            '6',
+            '60',
+            '61',
+            '62',
+            '63',
+            '64',
+            '65',
+            '66',
+            '67',
+            '68',
+            '69',
+            '7',
+            '70',
+            '71',
+            '72',
+            '73',
+            '74',
+            '75',
+            '76',
+            '77',
+            '78',
+            '79',
+            '8',
+            '80',
+            '81',
+            '82',
+            '83',
+            '84',
+            '85',
+            '86',
+            '87',
+            '88',
+            '89',
+            '9',
+            '90',
+            '91',
+            '92',
+            '93',
+            '94',
+            '95',
+            '96',
+            '97',
+            '98',
+            '99',
+        ];
+        const unorderedTags3 = createTagObjects(unorderedTagNames3);
+        const expectedOrder3 = createTagObjects(expectedOrderNames3);
+        expect(TagsOptionsListUtils.sortTags(unorderedTags3)).toStrictEqual(expectedOrder3);
+    });
+});
